requirements.txt
openai==0.27.8
mysql-connector-python==8.1.0
requests==2.31.0